@@ -11,32 +11,29 @@
         "start": "ng serve",
         "build": "ng build",
         "test": "ng test",
-        "e2e": "ng e2e",
-        "install:clean": "rm -rf node_modules/ && rm -rf package-lock.json && npm install",
-        "postinstall": "ngcc",
-        "eslint": "eslint --config .eslintrc.json --ext .ts,.html"
+        "e2e": "ng e2e"
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "10.2.2",
-        "@angular/cdk": "10.2.7",
-        "@angular/common": "10.2.2",
-        "@angular/compiler": "10.2.2",
-        "@angular/core": "10.2.2",
-        "@angular/fire": "^6.0.4",
-        "@angular/forms": "10.2.2",
-        "@angular/material": "10.2.7",
-        "@angular/platform-browser": "10.2.2",
-        "@angular/platform-browser-dynamic": "10.2.2",
-        "@angular/platform-server": "10.2.2",
-        "@angular/router": "10.2.2",
-        "@auth0/angular-jwt": "^5.0.1",
+        "@angular/animations": "11.0.2",
+        "@angular/cdk": "11.0.0",
+        "@angular/common": "11.0.2",
+        "@angular/compiler": "11.0.2",
+        "@angular/core": "11.0.2",
+        "@angular/fire": "^6.1.1",
+        "@angular/forms": "11.0.2",
+        "@angular/material": "11.0.0",
+        "@angular/platform-browser": "11.0.2",
+        "@angular/platform-browser-dynamic": "11.0.2",
+        "@angular/platform-server": "11.0.2",
+        "@angular/router": "11.0.2",
+        "@auth0/angular-jwt": "^5.0.2",
         "ag-grid-angular": "^24.1.0",
         "ag-grid-community": "^24.1.0",
         "ajv": "6.12.6",
         "angular-calendar": "0.28.22",
         "angular-gridster2": "^10.1.7",
-        "angular-svg-icon": "^10.0.0",
+        "angular-svg-icon": "^11.0.0",
         "animate.css": "4.1.1",
         "arrive": "2.4.1",
         "bootstrap": "4.5.3",
