@@ -1,17 +1,17 @@
-import {Component, OnInit, ElementRef} from '@angular/core';
-import {Router, NavigationEnd, NavigationStart} from '@angular/router';
-import {Subscription} from 'rxjs/Subscription';
-import {filter} from 'rxjs/operators';
+import { Component, OnInit, ElementRef } from '@angular/core';
+import { Router, NavigationEnd, NavigationStart, Event } from '@angular/router';
+import { Subscription } from 'rxjs/Subscription';
+import { filter } from 'rxjs/operators';
 
 @Component({
     selector: 'app-layout',
-    templateUrl: './public-layout.component.html'
+    templateUrl: './public-layout.component.html',
 })
 export class PublicLayoutComponent implements OnInit {
     private toggleButton: any;
     private sidebarVisible: boolean;
     mobile_menu_visible: any = 0;
-    private _router: Subscription;
+    private _router!: Subscription;
 
     constructor(private router: Router, private element: ElementRef) {
         this.sidebarVisible = false;
@@ -21,13 +21,15 @@ export class PublicLayoutComponent implements OnInit {
         const navbar: HTMLElement = this.element.nativeElement;
 
         this.toggleButton = navbar.getElementsByClassName('navbar-toggler')[0];
-        this._router = this.router.events.pipe(filter(event => event instanceof NavigationEnd)).subscribe((event: NavigationEnd) => {
-            this.sidebarClose();
-            const $layer = document.getElementsByClassName('close-layer')[0];
-            if ($layer) {
-                $layer.remove();
-            }
-        });
+        this._router = this.router.events
+            .pipe(filter((event: Event) => event instanceof NavigationEnd))
+            .subscribe((event: any) => {
+                this.sidebarClose();
+                const $layer = document.getElementsByClassName( 'close-layer' )[0];
+                if ($layer) {
+                    $layer.remove();
+                }
+            });
     }
 
     sidebarOpen() {
@@ -45,18 +47,22 @@ export class PublicLayoutComponent implements OnInit {
         var $layer = document.createElement('div');
         $layer.setAttribute('class', 'close-layer');
 
-
         if (body.querySelectorAll('.wrapper-full-page')) {
-            document.getElementsByClassName('wrapper-full-page')[0].appendChild($layer);
+            document
+                .getElementsByClassName('wrapper-full-page')[0]
+                .appendChild($layer);
         } else if (body.classList.contains('off-canvas-sidebar')) {
-            document.getElementsByClassName('wrapper-full-page')[0].appendChild($layer);
+            document
+                .getElementsByClassName('wrapper-full-page')[0]
+                .appendChild($layer);
         }
 
         setTimeout(function () {
             $layer.classList.add('visible');
         }, 100);
 
-        $layer.onclick = function () { //asign a function
+        // assign a function
+        $layer.onclick = () => {
             body.classList.remove('nav-open');
             this.mobile_menu_visible = 0;
             this.sidebarVisible = false;
@@ -66,12 +72,12 @@ export class PublicLayoutComponent implements OnInit {
                 $layer.remove();
                 $toggle.classList.remove('toggled');
             }, 400);
-        }.bind(this);
+        };
 
         body.classList.add('nav-open');
         this.mobile_menu_visible = 1;
         this.sidebarVisible = true;
-    };
+    }
 
     sidebarClose() {
         var $toggle = document.getElementsByClassName('navbar-toggler')[0];
@@ -93,7 +99,7 @@ export class PublicLayoutComponent implements OnInit {
         }, 400);
 
         this.mobile_menu_visible = 0;
-    };
+    }
 
     sidebarToggle() {
         if (this.sidebarVisible === false) {
