@@ -8,8 +8,8 @@ export enum PUBLISH_TYPE {
 }
 
 export class BindData implements Deserializable {
-    public dataId: string = '';
-    public queryParams: QueryParams[] = undefined;
+    public dataId!: string;
+    public queryParams!: QueryParams[];
 
     deserialize(input: any) {
         Object.assign(this, input);
@@ -18,8 +18,8 @@ export class BindData implements Deserializable {
 }
 
 export class BindVis implements Deserializable {
-    visId: string = '';
-    bindData: BindData[] = [];
+    visId!: string;
+    bindData!: BindData[];
 
     deserialize(input: any) {
         Object.assign(this, input);
@@ -28,12 +28,12 @@ export class BindVis implements Deserializable {
 }
 
 export class OntoPage implements Deserializable {
-    public id: string = '';
-    public title: string = '';
-    public bindVis: BindVis[] = undefined;
-    public nrows: number = undefined;
-    public publishType: PUBLISH_TYPE = undefined as any;
-    
+    public id!: string;
+    public title!: string;
+    public bindVis!: BindVis[];
+    public nrows!: number;
+    public publishType!: PUBLISH_TYPE;
+
     deserialize(input: any) {
         Object.assign(this, input);
         return this;
