@@ -26,9 +26,9 @@ export class QueryparamEditComponent extends BaseNestedform {
     
     // Remove tag
     remove(tag: any): void {
-        const index = this.formGroup.get('params').value.indexOf(tag);
-        if (index >= 0) this.formGroup.get('params').value.splice(index, 1);
-        this.nestedFormGroup.get('params').updateValueAndValidity();
+        const index = this.formGroup.get('params')?.value.indexOf(tag);
+        if (index >= 0) this.formGroup.get('params')?.value.splice(index, 1);
+        this.nestedFormGroup.get('params')?.updateValueAndValidity();
     }
 
     // Add tag
@@ -37,11 +37,11 @@ export class QueryparamEditComponent extends BaseNestedform {
         const value = event.value;
 
         // add tag
-        if ((value || '').trim()) this.formGroup.get('params').value.push(value.trim());
+        if ((value || '').trim()) this.formGroup.get('params')?.value.push(value.trim());
         // reset the input value
         if (input) input.value = '';
 
-        this.nestedFormGroup.get('params').updateValueAndValidity();
+        this.nestedFormGroup.get('params')?.updateValueAndValidity();
         // console.log('CollectionEditComponent: add: tags = ', this.collection.tags);
     }
 }
