@@ -3,7 +3,8 @@ import { InjectionToken } from '@angular/core';
 export const JQ_TOKEN = new InjectionToken('jQuery');
 
 export function jQueryFactory() {
-    return window['jQuery'];
+    const w: any = window;
+    return w['jQuery'];
 }
 
 export const JQUERY_PROVIDER = [{ provide: JQ_TOKEN, useFactory: jQueryFactory }];
